requirements.txt
# =============================================================
# NEXXI - Python Dependencies (pinned for reproducibility)
# =============================================================

# ── Core Web Framework ────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.1

# ── Hugging Face Ecosystem ────────────────────────────────────
transformers==4.41.0
huggingface-hub==0.23.0
accelerate==0.30.0
bitsandbytes==0.43.1
sentencepiece==0.2.0
# NOTE: Install torch separately with CUDA variant for GPU support:
# pip install torch==2.3.0+cu121 --index-url https://download.pytorch.org/whl/cu121
torch==2.3.0

# ── Conversation & Memory ─────────────────────────────────────
langchain==0.2.0
langchain-community==0.2.0

# ── Caching ───────────────────────────────────────────────────
redis==5.0.4
hiredis==2.3.2

# ── Security ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# ── Observability ─────────────────────────────────────────────
loguru==0.7.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
prometheus-fastapi-instrumentator==7.0.0
structlog==24.2.0

# ── HTTP Client (for integration tests) ──────────────────────
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# ── Code Quality ──────────────────────────────────────────────
black==24.4.2
ruff==0.4.4
mypy==1.10.0
